refactor: main como template
